Fix all remaining AUDIT.md medium/low findings (M-TR-4..L-4)

+++ av/metrics.go
@@ -429,7 +429,7 @@ func (ma *MetricsAggregator) updateSystemMetrics() {
 
 		totalPacketLoss += metrics.PacketLoss
 		totalJitter += metrics.Jitter
-		totalBitrate += uint64(metrics.AudioBitRate + metrics.VideoBitRate)
+		totalBitrate += uint64(metrics.AudioBitRate) + uint64(metrics.VideoBitRate)
 		totalDuration += metrics.CallDuration
 
 		// Count quality levels
